unify all the type classes into ToTH
It's becoming increasingly clear that I don't just need to reify
type-level values to the value-level, I must also convert them from the
GHC.Generics representation to the TemplateHaskell representation. So
this KnownFoo idiom isn't good enough. Furthermore, for KnownMaybeSymbol
I was disappointed to discover that I couldn't write a composable
KnownMaybe which would delegate to whatever KnownFoo is needed, I had to
specialize to KnownSymbol; with a single ToTH, I don't need to
specialize, because ToTH is the only choice.