move the work to ToTH, add Prime example
Now directly constructing Con's and Dec's instead of reifying the
type-level MetaData, MetaCons, and MetaSel values and then constructing
Con's and Dec's out of them.

Along the way, I added support for newtypes and records. I ignore the
isRecord boolean and instead check whether all the fields have names,
because if they don't then I can't create a record anyway.

The simple Tree2 example I had stopped working because I was using Rep
Tree as the source type, and so the output was now Tree, not Tree2. In
order to produce a Tree2, we now have to use type-generic programming to
modify Rep Tree, which is good, because that's how this library is
intended to be used. Since that's a lot more code, I moved it from a
doctest to a separate example file, and I made it slightly more
plausibly-useful by appending a prime instead of a two.